fix(groupChats): rename GroupChatsProgram
